refactor: separate functionality from the dashboard class
