+++ project-e-dukate-frontend/src/app/dashboard/especialidades/page.tsx
@@ -1,0 +1,5 @@
+import { Specialties } from '../../../pages/Specialties';
+
+export default function SpecialtiesPage() {
+  return <Specialties />;
+}
