+++ project-e-dukate-frontend/src/app/dashboard/horarios/page.tsx
@@ -1,0 +1,5 @@
+import { Schedules } from '../../../pages/Schedules';
+
+export default function SchedulesPage() {
+  return <Schedules />;
+}
