+++ project-e-dukate-frontend/src/app/dashboard/metricas/page.tsx
@@ -1,0 +1,5 @@
+import { Metrics } from '../../../pages/Metrics';
+
+export default function MetricsPage() {
+  return <Metrics />;
+}
