+++ project-e-dukate-frontend/src/app/dashboard/pagos/page.tsx
@@ -1,0 +1,5 @@
+import { Payments } from '../../../pages/Payments';
+
+export default function PaymentsPage() {
+  return <Payments />;
+}
