+++ project-e-dukate-frontend/src/app/dashboard/usuarios/page.tsx
@@ -1,0 +1,5 @@
+import { Users } from '../../../pages/Users';
+
+export default function UsersPage() {
+  return <Users />;
+}
