+++ project-e-dukate-frontend/src/components/Sidebar/Sidebar.tsx
@@ -16,19 +16,18 @@ interface MenuItem {
 
 interface SidebarProps {
   selectedTab: string;
-  onTabChange: (tab: string) => void;
   sx?: SxProps;
 }
 
-export const Sidebar: React.FC<SidebarProps> = ({ selectedTab, onTabChange, sx }) => {
+export const Sidebar: React.FC<SidebarProps> = ({ selectedTab, sx }) => {
   const router = useRouter();
 
   const menuItems: MenuItem[] = [
-    { label: 'Especialidades', icon: <FaStethoscope size={20} />, value: 'especialidades' },
-    { label: 'Usuarios', icon: <PeopleAltOutlinedIcon />, value: 'usuarios' },
-    { label: 'Pagos', icon: <PaymentsOutlinedIcon />, value: 'pagos' },
-    { label: 'Horarios', icon: <CalendarMonthOutlinedIcon />, value: 'horarios' },
-    { label: 'Métricas', icon: <BarChartIcon />, value: 'metricas' },
+    { label: 'Specialties', icon: <FaStethoscope size={20} />, value: 'especialidades' },
+    { label: 'Users', icon: <PeopleAltOutlinedIcon />, value: 'usuarios' },
+    { label: 'Payments', icon: <PaymentsOutlinedIcon />, value: 'pagos' },
+    { label: 'Schedules', icon: <CalendarMonthOutlinedIcon />, value: 'horarios' },
+    { label: 'Metrics', icon: <BarChartIcon />, value: 'metricas' },
   ];
 
   const handleLogout = () => router.push('/login');
@@ -43,7 +42,7 @@ export const Sidebar: React.FC<SidebarProps> = ({ selectedTab, onTabChange, sx }
           {menuItems.map((item) => (
             <ListItemButton
               key={item.value}
-              onClick={() => onTabChange(item.value)}
+              onClick={() => router.push(`/dashboard/${item.value}`)}
               sx={{
                 mx: 2,
                 borderRadius: 2,
