+++ project-e-dukate-frontend/src/pages/Dashboard/index.ts
@@ -1,1 +1,1 @@
-export { default as Dashboard }  from './Dashboard';
+// export { default as Dashboard }  from './Dashboard';
