+++ project-e-dukate-frontend/src/pages/Metrics/index.ts
@@ -1,0 +1,1 @@
+export { default as Metrics } from './Metrics';
