+++ project-e-dukate-frontend/src/pages/Payments/index.ts
@@ -1,0 +1,1 @@
+export { default as Payments } from './Payments';
