+++ project-e-dukate-frontend/src/pages/Schedules/index.ts
@@ -1,0 +1,1 @@
+export { default as Schedules } from './Schedules';
