+++ project-e-dukate-frontend/src/pages/Specialties/index.ts
@@ -1,0 +1,1 @@
+export { default as Specialties } from './Specialties';
