+++ project-e-dukate-frontend/src/pages/Users/index.ts
@@ -1,0 +1,1 @@
+export { default as Users } from './Users';
